refactor: Move moneyutils to pkg
